pyproject.toml
[project]
name = "cakto-split-payment-challenge"
version = "0.1.0"
description = ""
authors = [
    {name = "lucascbenedet",email = "lucasbenedet8@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "prometheus-client (>=0.22.1,<0.23.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false